backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
duckdb>=1.1.3
anthropic==0.42.0
httpx>=0.27.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
PyJWT>=2.8.0
bcrypt>=4.1.0
